please read remarks in description of this commit
Large refactor of the system description.
Below I have listed, in no particular order, some remarks/tips.
I think the most important remark is #3.

1) There was inconsistent use in singular/plural for the key object
classes, I made it singular everywhere.
2) Please enter newlines after each sentence in tex, it makes both the
tex file and the diff so much more readable.
3) Please don't write a report like it's a diary.
4) Please make text italic by surrounding it with \textit{} in stead of
$$, this prevents strange constructions like $Waiting \ for \ Cashier$
but it also looks better for single words especially when they contain
an f.
5) Please use \setlength{\parskip}{<length>} and
\setlength{\parindent}{<length>} instead of endlines and \indent
statements
6) er stond in het verslag dat elke line 2 pumps heeft, maar volgens mij
klopt dat niet: "To stream the traffic, the fuel station will have 15
fuel pumps, divided in 5 groups. Cars
queue in one of the 15 lanes and wait for their turn at the pump."
Furthermore: "Each of the 15 pumps allow for
both low and high-speed operation depending on the vehicle type."
So I think the type distinction for the pumps can be omitted .
This raises the question: Do we need the class pump at all? or can we
handle the availability of the pump as a special spot in the lane
somehow?